If you are applying this as a patch to your running mud, be sure to
                         BACK UP YOUR FILES FIRST

This is especially true, since the circlemud 3.5 code released
during december 2006 WILL RUIN YOUR HOUSE FILES and CORRUPT your
rent files. This is due to the introduction of ASCII rent files.
 
Unfortunately we were too eager to add what we thought was a
magnificent idea, and in the process forgot that people with
actual live running muds might use the code we provided.

Since the development has taken place on TBA, which has no players
as such, we failed to notice that the update was not thorough enough,
and several important additions were not added - for instance, a
house file and rent file converter tool. Also, it turned out, the
addition had severely hampered the object saving and loading routines,
to a point where items were being saved in an incompatible format
in houses.

Here is how to successfully convert your circlemud 3.1 house files,
step by step:

1) backup your house files somewhere not in the /lib/house folder
2) apply patch, debug, etc.
3) run the mud
4) now copy your house files back (while the mud is running!)
5) from within the mud, type "hcontrol asciiconvert" (without the quotes).
   This will give you some feedback about which houses are being processed.
6) shut down the mud
7) in lib/house, enter these commands:
    rm *.house            (you still have a backup of this, right ?)
    for i in *.ascii; do mv "$i" "${i/.ascii}"; done
8) start the mud - your houses now contain the right items.
9) feel free to remove code in house.c marked with CONVERSION tags.