Standardize on American spellings.
Full internationalization would be much cooler, but that's never going
to happen.  Given that, this will at least prevent folks from constantly
having to look and switch between typing "armor" and "armour", depending
on which zone each item originated in, etc.

I could flip these either way, but a survey of the current state shows
that about 80% of the mixed cases use the American spellings, while 20%
use the British.  And, most words *only* exist in this data in their
American forms.  So, it seems the majority prefer these spellings.

In case anyone likes trivia:
* The most common mixed words in here were "armour" and "colour", each of
  which occured about half as often as "armor" and "color", respectively.
* The most British word in here was "theatre" (including other forms),
  which occured about twice as often as "theater".

This stanardizes all of these (and other forms of these same words):
* armour -> armor
* colour -> color
* favour -> favor
* honour -> honor
* civilise -> civilize
* centre -> center
* theatre -> theater
* defence -> defense
* offence -> offense
* realise -> realize