Update handler.c
IN:
void affect_total()

REPLACED:
  /*Make certain values are between 0..25, not < 0 and not > 25! */
  i = (IS_NPC(ch) || GET_LEVEL(ch) >= LVL_GRGOD) ? 25 : 18;

WITH:
  /*Make certain values are between 0..LVL_IMPL, not < 0 and not > LVL_IMPL! */
  i = (IS_NPC(ch) || GET_LEVEL(ch) >= LVL_GRGOD) ? LVL_IMPL : 18;

RATIONALE:
Old code sets real_abils to a max of 25. Yet, we have 34 levels. Why should real_abils be limited to 25 for >= LVL_IMMORT?

Maximum real_abils values should always be equal to LVL_IMPL. This permits a maximum increase of +1 for every level. It allows infinite level expansion, for coders that wish to add more levels, with continued increases in real_abils, and with fewer code changes.