Update fight.c
REPLACED:
armorclass += dex_app[GET_DEX(ch)].defensive * 10;

WITH:
armorclass += dex_app[GET_DEX(ch)].defensive * 1;

RATIONALE:
Old code was designed to operate within a (+10) to (-10) range. That was fine for the old dex_app[] entries and values. However, the d20 standardized dex_app[] is intended to operate with in a (+10) to (-100) range for AC. The old code will function fine with the new dex_app[] , but players will max AC at level 20.

Total number of entries in dex_app[] should always be equal to LVL_IMPL. This permits a a maximum increase of +1 (or -1 to AC) for every level. This allows an increase in levels for coders that wish to add up to 100 levels, with continued increases for AC, while maintaining game balance. Furthermore, this moves tbaMUD closer to a more modern d20 system.